requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
black==20.8b1
boto3==1.17.44
botocore==1.20.44
certifi==2020.12.5
click==7.1.2
dj-database-url==0.5.0
Django==3.1.7
django-modeltranslation==0.17.1
django-phonenumber-field==5.0.0
django-storages==1.11.1
gunicorn==20.0.4
jmespath==0.10.0
mypy-extensions==0.4.3
numpy==1.20.2
pandas==1.2.3
pathspec==0.8.1
phonenumbers==8.12.20
psycopg2-binary==2.8.6
pycodestyle==2.6.0
python-dateutil==2.8.1
python-telegram-bot==13.3
pytz==2021.1
redis==3.5.3
redis-persistence==0.0.4
regex==2021.4.4
s3transfer==0.3.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
XlsxWriter==1.3.8
